requirements.txt
# BharatChef AI Coach - Python Dependencies
# Downloaded from working EC2 instance

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# AWS SDK
boto3==1.29.7
botocore==1.32.7

# Task Queue
celery==5.3.4
redis==5.0.1

# Environment Variables
python-dotenv==1.0.0

# Data Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Utilities
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dateutil==2.8.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Computer Vision Dependencies
torch==2.1.0
torchvision==0.16.0
opencv-python==4.8.1.78
opencv-python-headless==4.8.1.78
numpy==1.24.3
pillow==10.1.0
